I went to bed for a bit to think. Right now it really bothers me that tanh plateaus with low learning rates. That should not be happening. It really bothers me deeply.
The is the main thing I need to check out. It might be good for me to go a month or two back and run those tests again. Was I fooled by randomness or did I get something deeply wrong in the latest tests.

The other thing I need to check out is the math for PG's backward pass. Should it really be `(label - input) * reward`?

Lastly is the issue of critics. Yeah, I knew I said I would do the critics and am just sitting here instead of programming, but I can also think of several ways of doing this and am not sure where to start.

Thankfully after deliberating on this for a while I realized that the situation is simpler than it seems at first glance. There are a bunch of scenarios, but since TD learning is unstable with function approximation, what I should do is just implement Zap and not even bother with the vanilla versions.

If I want stability, I suppose I could train a critic using MC, but forget that stuff. I want to get away from that.

Forget it. It is Zap and and pretty much nothing else. I won't explore the lesser possibilities and I won't concern myself too much with this.

Now with that out of the way, I need to figure out the reason for the recent underperformance.

I hate this sort of debugging as it could turn out to be anything. Unlike last time I am not scaling the rewards, but learning rates should take care of that.

I am going to have to explore it a bit. Let me see if I can deal with this today.