Yeah, the weights are equal. Now let me move to the next step. I will control for the random seeds on the .NET side. If I do that then the runs should be fully reproducible.
```
Starting timing for: iteration 0
Winrate is 212 and 788 out of 1000.
The time was 00:00:04.1844941 for: iteration 0
Starting timing for: iteration 1
Winrate is 344 and 656 out of 1000.
The time was 00:00:04.6171866 for: iteration 1
Starting timing for: iteration 2
Winrate is 461 and 539 out of 1000.
The time was 00:00:04.0089991 for: iteration 2
Starting timing for: iteration 3
Winrate is 557 and 443 out of 1000.
The time was 00:00:04.1985380 for: iteration 3
Starting timing for: iteration 4
Winrate is 574 and 426 out of 1000.
The time was 00:00:04.0363119 for: iteration 4
```
...
```
Winrate is 212 and 788 out of 1000.
The time was 00:00:04.0785667 for: iteration 0
Starting timing for: iteration 1
Winrate is 344 and 656 out of 1000.
The time was 00:00:04.5558845 for: iteration 1
Starting timing for: iteration 2
Winrate is 461 and 539 out of 1000.
The time was 00:00:04.0265579 for: iteration 2
Starting timing for: iteration 3
Winrate is 557 and 443 out of 1000.
The time was 00:00:04.1213147 for: iteration 3
Starting timing for: iteration 4
Winrate is 574 and 426 out of 1000.
The time was 00:00:03.9719328 for: iteration 4
```

Now that I am controlling for the seeds the runs in the old version are fully reproducible. Let me do the same for the new thing and then I will call it a day. I'll debug this to completion tomorrow. It will probably turn out that the serialization functions are messing up somewhere. It might even be in the old library.