"7:10pm. Now the two games seem to be the same.
```
[|
    [|0; 0; 0; 1; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 1; 0; 0; 0; 0; 1; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 1; 0; 0; 0; 0; 0; 0; 1; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 1; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0|]
|]

[Raise, 0]
```
...
```
[|
    [|0; 0; 0; 1; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 1; 0; 0; 0; 0; 1; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 1; 0; 0; 0; 0; 0; 0; 1; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 1; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0|]
|]

[Raise, 0]
```

Yeah, they are the same. Let me check out the winrates. I'll remove the print functions.

7:10pm.

```
Starting timing for: iteration 0
Winrate is 242 and 758 out of 1000.
The time was 00:00:03.9665240 for: iteration 0
Starting timing for: iteration 1
Winrate is 402 and 598 out of 1000.
The time was 00:00:03.7243304 for: iteration 1
Starting timing for: iteration 2
Winrate is 458 and 542 out of 1000.
The time was 00:00:03.8387637 for: iteration 2
Starting timing for: iteration 3
Winrate is 530 and 470 out of 1000.
The time was 00:00:03.9782459 for: iteration 3
Starting timing for: iteration 4
Winrate is 557 and 443 out of 1000.
The time was 00:00:03.9829717 for: iteration 4
Starting timing for: iteration 5
Winrate is 649 and 351 out of 1000.
The time was 00:00:04.2945477 for: iteration 5
Starting timing for: iteration 6
Winrate is 655 and 345 out of 1000.
The time was 00:00:04.3130372 for: iteration 6
Starting timing for: iteration 7
Winrate is 679 and 321 out of 1000.
The time was 00:00:04.3631699 for: iteration 7
Starting timing for: iteration 8
Winrate is 686 and 314 out of 1000.
The time was 00:00:04.0319672 for: iteration 8
Starting timing for: iteration 9
Winrate is 680 and 320 out of 1000.
The time was 00:00:03.9247557 for: iteration 9
Starting timing for: iteration 10
Winrate is 715 and 285 out of 1000.
The time was 00:00:03.9125064 for: iteration 10
Starting timing for: iteration 11
Winrate is 698 and 302 out of 1000.
The time was 00:00:03.8396427 for: iteration 11
Starting timing for: iteration 12
Winrate is 726 and 274 out of 1000.
The time was 00:00:03.5100408 for: iteration 12
Starting timing for: iteration 13
Winrate is 697 and 303 out of 1000.
The time was 00:00:03.5733004 for: iteration 13
Starting timing for: iteration 14
Winrate is 712 and 288 out of 1000.
The time was 00:00:03.5327296 for: iteration 14
```

Hmmm...the old one seems to work well regardless. Hmmm, what the hell?

Let me compare it with the new one.

```
Starting timing for: iteration 0
Winrate is 242 and 758 out of 1000.
The time was 00:00:04.1101484 for: iteration 0
Starting timing for: iteration 1
Winrate is 402 and 598 out of 1000.
The time was 00:00:03.7967596 for: iteration 1
Starting timing for: iteration 2
Winrate is 458 and 542 out of 1000.
The time was 00:00:03.9841156 for: iteration 2
Starting timing for: iteration 3
Winrate is 530 and 470 out of 1000.
The time was 00:00:04.1516139 for: iteration 3
Starting timing for: iteration 4
Winrate is 557 and 443 out of 1000.
The time was 00:00:04.0684851 for: iteration 4
Starting timing for: iteration 5
Winrate is 649 and 351 out of 1000.
The time was 00:00:04.4241170 for: iteration 5
Starting timing for: iteration 6
Winrate is 655 and 345 out of 1000.
The time was 00:00:04.4324424 for: iteration 6
Starting timing for: iteration 7
Winrate is 679 and 321 out of 1000.
The time was 00:00:04.4795629 for: iteration 7
Starting timing for: iteration 8
Winrate is 686 and 314 out of 1000.
The time was 00:00:04.1744811 for: iteration 8
Starting timing for: iteration 9
Winrate is 680 and 320 out of 1000.
The time was 00:00:03.9692144 for: iteration 9
Starting timing for: iteration 10
Winrate is 715 and 285 out of 1000.
The time was 00:00:03.9952758 for: iteration 10
Starting timing for: iteration 11
Winrate is 698 and 302 out of 1000.
The time was 00:00:04.0261411 for: iteration 11
Starting timing for: iteration 12
Winrate is 726 and 274 out of 1000.
The time was 00:00:03.5474828 for: iteration 12
Starting timing for: iteration 13
Winrate is 697 and 303 out of 1000.
The time was 00:00:03.6411709 for: iteration 13
Starting timing for: iteration 14
Winrate is 712 and 288 out of 1000.
The time was 00:00:03.6081743 for: iteration 14
```

Yeah, the new one works well now too.

7:15pm. Uagh, I was afraid of this. I think the real lesson might be that I should randomize the hyperparameters for weight init.

This is such bullshit. I'll need to look into why the change in representation resulted in such drastic changes in trajectory, but I'll leave that for tomorrow. It is most likely randomness. I'll have to randomnly init the weights and see exactly how much the runs differ. Probably the fact that I've been using the same nets everywhere is biasing my judgement of various methods.

Right now it is time for play and rest. I put in some serious overtime today.

But it is good to go through experiences such as these from time to time."