It is time for me to start the programming day. I am bored of everything else, so this is the ideal frame of mind to focus on what I actually have to do.
Yesterday I did a lot of work that turned out to be pointless. The underporformance of tanh in the new library is not due to some bug, but merely because of randomness. So today what I am going to do is randomize the CudaRandom seeds in order to get a more accurate picture of how it performs in practice.

Yesterday I had some time to think about the critics, but if I focus on TD learning style critics specifically then Zap in the last layer is really the only option. Considering linear layers blow up on Baird's counterexample, that is absolute proof that RL needs some kind of variance reduction sheme, and since Zap is specialized for its role, I will be better off chosing it rather than the more generic PRONG.

Apart from Zap, really the only option from a stability standpoint would be vanilla layers - deep nets are out of the question as they will just explode if the Q learning experiment was anything to go by.

Apart from AC, I really should implement Zap Q as well and use it as my Q learning baseline.  Poker has a small enough number of inputs that this might just be doable.

Actually, I could use it straightforwardly too in a deep net. Gradients could be propagated in a standard manner by treating it as a linear layer.

...I should definitely try this. Deep Zap Q vs Deep Zap AC would be an interesting battle to be sure.

At this point in my journey it is absolutely imperative that I attain roboustness across a wide range of starting weights. PRONG and Zap will be the tools to attain this.