While this is running, let me go take a break. I need to see just how lucky or unlucky I really am with the tanh.
I'll bring in PRONG and LN later.

Well, there will be plenty of time for testing. I'll go through all of it in turn. This will take me a while.

Before I do all of that, I think I will start work on Zap Q and Zap TD today. I want to get this phase over with. Interfacing the agents with the world is something that will bring me great profit.

What I am doing here is just the prelude to that.