I figured out some things.
I started thinking about gradient flows and realized that AC has two levels of dampening that Q learning specifically lacks.

1) `(p - label) * reward`. This is how it is done on the backward pass, meaning that as the distribution approaches certainty, the policy moves will be less pronounced. Q learning does not have this on the top level.

2) The second level would be the weights. In Q learning, the weights are used to calculate the values directly. So they will reflect the scale of the values and this will cause the gradients to be pushed twice in that direction. In other words, if the rewards are big, then those big rewards will also be multiplied by large weights on the backward pass. On the other hand if the rewards are small, they will be multiplied by small weights on the backward pass. This is what I mean by gradients being pushed twice in the direction of the scale of the reward.

In AC like in PG, the weights in the last layer will be in probability space rather than reward space which is much better.

These two things could actually explain why PG methods are much more stable than Q learning.

I've been wondering for the entire day whether I should let gradients flow through the critic and the answer is - absolutely not. This also gives me an answer to the question of whether Zap would be enough to stabilize Q learning. It would not.

Add to that the benefit of exploration that AC has and it is the clear choice here. My original plan was more sound than I imagined.