https://www.reddit.com/r/reinforcementlearning/comments/959ok6/d_should_gradients_be_propagated_through_the/
That takes care of that. Now that I have this out of my system what is next?

For matrix inversion of regular matrices I am thinking of introducing the LU decomposition.

http://www2.cs.cas.cz/semincm/lectures/2010-07-1920-DuintjerTebbens.pdf

"A survey of properties of methods for matrix inversion based on triangular decompositions is given in [Du Croz, Higham - 1992]. In the following we summarize the main results on implementation and stability. We start with the inversion of a unit lower triangular matrix L."

Oh, this is exactly what I needed to make the decision of whether to do it. Let me read these slides.