Let me resume programming. Considering my goal it to do the matrix inverse using LUdecomposition, I am not sure why it is taking me this long.
Based on the paper it does seem like A3C works even without target networks.

Uah...just forget it. I won't speculate on this anymore. This is such an amazing pain in the ass. It really goes to show that the RL community will make anything work if they throw enough compute at it. I wish they were more thorough in documenting the effect of various architectural choices.

But my insight is not wrong. It is true that for Q learning the gradients are in the square reward space.

Let me bet on that insight. I need to focus and implement the matrix inverse so I can get to Zap. I cannot allow this to hold me down.

It needs to be done today, period.