This really took a lot of work to get right. I knew it would. Messing with dimensions always brings trouble with it.
I am not done with it yet. I need to figure out how to decompose non-square matrices. Should I just leave that for tomorrow?

Well, let me give it a shot for just a bit. It is true that I am near my limit, but I still have some momentum left in me. I'll leave the actual inverse for tomorrow.