Let me get rid of that stuff in the previous commit so I can put the assert back in.
I'll dedicate tomorrow to brining in the Cuda callbacks. After that it will be very easy to implement Zap.

When I bring back PRONG one thing I need to make sure is to compare it with inverse that uses potrs instead of those triangular matrix multiplies.

Well, I'll take it easy for the rest of the day. It is time to rest.