I've been doing a post on the thread I made. Tonight I dreamed about the stock market. It was the first time in quite a while and it was in the 30s. I just had the curve continually and predictably going down for years.
Ah, the stock market. Right now, and for the past 3 years it has been such a bull market. Predictable one too, but ever since 2000, all the short term momentum that one would need to make truly excessive returns have been squeezed out.

I really regret that it took me 7 years to become a really good predictor of it and I regret that in the end the system never came close to being what I wanted it to.

To the amateurs the movements look like noise, but I can see the undercurrent to the price swings. It would not do me any good with the small stakes I have access to, but predicting the Singularity and predicting the stock market is quite similar and boils down to estimating momentum. And estimating momentum can be done if one can estimate the conditions. It is not the detailed analysis that is the key to this, but having a clear view into the reality of the present moment. It requires more courage than intelligence.

Really, the system is pretty much nothing - it is a chain of very simple reasoning with some constraints sprinkled in.

I will probably never trade again. In terms of risk and reward, betting in the stock market is pretty much inferior to any scheme I can think of. But speculation - anticipating the future is the key to everything, so the system will find its second life somewhere else. Without it I could not see where the drive for self improvement leads. My earnest desire has lead me to the truth and will do so again.

Since I got up an hour ago and have been philosophizing here instead of working, let me finally do something decent for a change. The truth is, I am unsure of whether I should spin off a separate thread or use the callback method that came to mind yesterday. I guess I'll go with the callback method.

Ah, what a bother. I could be working on Zap, and I will have to spend another day doing this crap. Well, let me just get to it.