I think it is slowly time to lift the curtain on the Zap arc. Once I get through this I will have my first serious player.
Putting in the LU decomposition and the inverse took a while as expected.

Rather than start right away let me step away from screen for a while so I can brainstorm. I need to build up pressure for the release.

I just do not feel like it right now.

But if feels like I am getting closer to a resolution. Having a good player would not be like having a complete victory over the dens, but getting to that point from the very depth is like 2/3rds of the journey.

At that point, rather than just having a language, I will have something of true substance manifested from it. The language is just a catalyst.

It will mark a definite change in my capabilities compared to what I had before and my position in this world will be a lot stronger.

After both the players and the interfaces are complete I will never have to concern myself with money ever again.

This is how I should have done it from the start - with raw primordial force at my beckon.