Did some cleaning up of the inverse functions so they do not clutter up the tests.
Actually, it is pretty difficult for me to make the update function in isolation. The Cuda project is the wrong project to do it in. It needs to be in the Game project.

I am going to feed it actual states and actions, that will get me the fastest where I want to go.

Since it is 5:20pm, I am not sure how much I'll manage to do, but let me see if I can do the scaffolding at least. Right now I feel in the zone so let me do some more work.

I am thinking that for exploration once I am done I will optimistically init it. I do not think I should do this for deep nets - for those I should init the weights to zero, since I will have the actor doing the exploration there, but for Zap Q I think optimistic init should do nicely.