I need to stop here for a bit to think. I've been doing that for a while.
There is a conflict in how I want to do things and with how Zap Q is designed, though I might be able to take advantage of this situation as an opportunity.

Simply put, I've been intending to use backprop through time with recurrent nets. This would necessitate backward chaining of the rewards.

Zap Q, fortunately or unfortunately uses eligibility traces and does not require this.

Not a problem with feedforward nets, but with recurrent ones I am not sure. There have been papers on some architectures that do not require BPTT, but I haven't been paying them close attention.

But I have been thinking about this.

I do not think I should use backward chaining with Zap Q. Instead I should take advantage of its strengths.

Yeah, I think for the time being why don't I just focus on feedforward nets? After I push them to their limit how about I move to bigger and better things after that?

That having said, why don't I stop here for the day. Making this decision means abandoning my earlier approach and I need to come to terms with it.

I also want to take some time off in order to research BPTT alternatives. Actually I really want to drop BPTT as it won't be scalable at all in the longer term. The ideal time to do it would be as soon as possible.

I saw some papers on UORO and follow ups on it, but I never really bothered properly understanding them so I am not sure whether they are intrinsic to particular architectures.

...Actually no, am I stupid? I forgot something really really obvious about eligibility traces.

I can just set the eligibility_factor(lamda) hyperparameter to 0. That would default the eligibility trace to acting as the state.

This would simplify the code quite a bit and would give me back my backward chaining.

...Ah, I see it now. Yeah, the TD(0) version of the algorithm is much more understandable. I definitely made a step in figuring out Zap.

I need to redo all of this. Eligibility traces should definitely not be a part of the network anyway. I was really confused in that part. They should be passed in as the part of the forward fold. The intended use case for TD(lambda) should be the forward fold and for TD(0) the backward fold.

Not now though. I think I did more than enough for the day. I've been too enthusiastic yesterday when I said I would be done with this today. Before I come to a decision on what I want to do let me do research on BPTT alternatives. At the very least I want to make sure that I can adapt those things so they can be used in tandem with differentiable plasticity.