I made some changes. First I introduced the {from to} arguments in copy, cholesky_inverse and lu_inverse that slipped my mind earlier.
In addition to that I also removed the division by batch size and replaced those map and reduce operations in update_weights with much simpler matrix multiplies.

Now the whole thing is perfection. I need to run the tests for everything again though.

...Now, most of the code in the zap_q player should be factored out and it will be as I will need to use it to implement a Zap Critic, but for now I am fine with it being where it is for now.

I was going to stake a break, but instead started implementing things as soon as I got up so let me take that break now.

After that is done, I'll rerun all the Cuda tests and then move to testing the new Zap player. I have great expectations of it in terms of its stability and convergence, but it might turn out to be that case exploration will be a crucial bottleneck..