No matter how I slice it it not enough for me. Even if Zap meets my performance requirements, in order to really graduate and move onto the next phase of my plan, I need to properly link it.
What Rudder does and what standard deep RL does are inverses of each other.

If Zap is really doing what I think it is doing then it should be enough to obsolete having the need to do explicit reward decomposition.

Rudder is probably one of most important papers in deep RL ever, but it is the kind of idea that should have been obvious 30 years ago. Because the timeline is messed up the field will need some catching up to do.

Rudder's true purpose seems to me more like an object of assault. It is the ideal which the other methods need to strive to reach rather than being used directly.

If I want to prove that I understand the basics of RL, I need to understand Zap, use it unbias the recurrent network gradients and obsolete BPTT, and prove that it does optimal reward decomposition.

Obsoleting BPTT has been on my mind for a while. Those papers I've read are garbage - there is no doubt in my mind that the solution to this will be an update like Zap. PRONG does not seem suited to this, but it should be possible to do better.

The reason why I am mentioning Zap in this context is because all RL nets are in fact recurrent. This is a fact that is not appreciated right now, but should become obvious to the community at some point.

Since Zap's reprojection step has an unbiasing effect on the gradients it is reasonable to wonder whether it could be applied more generally.

Break time is over. Let me run the Cuda tests from the top and then I will get serious.

The timeline to start the Ascent would be 3.5 years. That is a long enough time for me to grasp the basics and hit the ground running.

Once we are past the basics, the world will move into a new regime where individual self improvement aptitudes dominate.