``` ------ The CudaRandom pseudorandom seed is 0 Starting timing for: iteration 0 Winrate is 261 and 739 out of 1000. The time was 00:00:00.9607100 for: iteration 0 Starting timing for: iteration 1 Winrate is 280 and 720 out of 1000. The time was 00:00:00.7456442 for: iteration 1 Starting timing for: iteration 2 Winrate is 316 and 684 out of 1000. The time was 00:00:00.7639551 for: iteration 2 Starting timing for: iteration 3 Winrate is 284 and 716 out of 1000. The time was 00:00:00.7708263 for: iteration 3 Starting timing for: iteration 4 Winrate is 307 and 693 out of 1000. The time was 00:00:00.7650020 for: iteration 4 ```
I am convinced the Zap is doing something remarkable. It is certainly preventing the network from diverging. Without the steady state matrix the linear network diverges 100% of the time. Still I can't get the winrate to go above 30% for some reason. I've gone from using optimistic init to epsilon greedy for exploration. For the simple game I am using here it should suffice.

Definitely there is an error somewhere, but I am not sure where exactly.

Well, this is a normal part of making ML algorithms work. I'll try printing things out and making an in depth investigation of this. I need to get this to work before I can go to AC anyway. It needs to be spotless.