"4:30pm.
```
    [|252.705; 1621.857; -306.8287; 36.31404; -1127.752; 87.85537; 419.9361; 312.9087; 100.1606; 62.09849; -1180.751; 54.58205; 58.54973; -205.0618; 320.8797; -13.18362; 56.51606; 0; 0; 0; 0; 87.8196; -141.6862; 6.311172; -138.0036; -122.2833; 4.567605; -43.63186; -25.21367; 14.08006; -73.0243; 39.22472; -79.45063; 18.76383; 49.90752; 0; -32.28632; 308.4861; -334.7793; 201.5181; 123.3677; 2079.634; -97.83486; -560.2746; -491.7951; -377.1188; 245.2478; 0; 0; 0; 0; 0; 0; 0; 0; 0; 37.09853; -1246.228; 30.88427; -1720.478; 2987.788; 192.2657; 747.9747; -44.72432; 149.8445; -614.3338; 2117.137; -1361.822; 55.90004; -4908.363; 156.7761; 442.4529; -1458.045; 0; 0; 0; 0; 62.95364; 102.4358; -551.8045; -274.7261; -232.8693; -194.5343; -191.4702; 552.7449; -64.24554; 380.215; -100.733; -2.742649; -152.4933; -705.597; 0; 0; 106.3858; -12.48663; -1089.719; -51.46386; -264.7513; -2900.634; -567.4672; -264.5555; 293.1993; 0; 0; 0; 0; 0; 0; 0; 0; 0; 0; -575.9549; -605.0948; 2526.247; 3407.3; -882.8121|]
```

Well, I am starting to understand why the steady state matrix causes the training to blow up. How could this be possible? The regularization should prevent this from happening.

I do not think I should concern myself about the errors during normal iteration. What is going on is that the network is making mostly random moves so it gives the appearance of things going wrong.

```
inl identity_coef = 0.2f32
```

Let me increase the regularization.

```
                cd.CudaKernel.iter {dim=A.dim} <| inl a b ->
                    if b <= a then
                        inl A = A a b
                        inl identity = if a = b then epsilon else zero
                        A .set (alpha * A .get + identity)
```

Ahhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhh....note that `if b <= a then` there. That definitely should not be there as the update is not symmetric here. This is what copy paste gets me.

4:45pm. Ok, I can get the last step to not blow up together when the steady state matrix is used.

There is some hope again. Let me try running the full thing once more."