Ok, I've ascertained that the update to the steady state matrix's bug is the culprit of all instability.
Now that I've fixed it the agent gets a winrate of about 20%. This is absurd.

I do not think this is because of exploration either.

There has to be some kind of other error lurking somewhere.

Well, there are some memory bugs, but those would only cause fragmentation, not outright permance divergence.

Let me print out W_action just so I can see what is happening.

I'll give it some more thought the usual way, and if that does not work, I am going to make a different simplified test to get to the bottom of this.

Maybe for that tabular example I did TD(0) on.