Yesterday I went to bed really early as I was so bored and I wanted to started programming as soon as possible.
During the night I figured out what the problem with the code is. I did something really really stupid with the basis function. It is the kind of error that should have been obvious immediately to me before I even started work on this, but I missed it somehow.

Plainly put, by having the action values be single features they just act as biases across all the states. Meaning a value of a action will be the same regardless of the state and will just act as an average.

Crap. I need to redo it all over again.

Honestly, I think I will just dive into the actor critic. I do not want to chase Zap Q anymore than this.

The reason why I wanted the Q version is because I thought it would make isolating the improvement to the critic easier, but this is not going well in that respect.

Now that I know where my mistake is I know how to do the Zap Q version, but the basis fuctions for that are would involve a lot of complications and would cause the dimensions to scale multiplicatively by the number of actions.

Actually implementing AC which I wanted from the start is a lot easier for me.