I got distracted for a bit by chores, but regardless I need to stop here.
This is hard, really really hard. Both the pg and the zap_ac are of low quality code. All the boxing and type inference that pg does is making it really difficult for me to merge the two together. I cannot paste the two together, because all the hand inling has crossed the threshold of my brain is capable of grasping.

This is unacceptable. I am not sure how I am going to do this, but I need to sit down a simplify the PG player and factor out all that is possible. The same goes for Zap.

I've been going fast way too much. For today, I am going to focus just on design work. After I am done with that, the player will pretty much finish itself.