I need a break at this point. All this thinking is starting to make me nauseous. It is like trying to make my way through a maze.
I am going to change the semantics of the language. I'll allow the case operation to `indiv` the damn variables. This is just really really annoying right now. I really was not counting on individual branches being layout types. Needing to do this by hand would simply be too much.