I am going to change the semantics of the language in order to allow case to indiv the layout types. This was the previous behavior and then I changed it to the current more restrictive one. It was a mistake after all.
Right now it is forcing me to go through all sorts of convoluted steps when layout types get used in tandem with union types.

It is just too difficult to keep track of when exactly a type should be unpacked, but it is easy enough to keep track when it should be packed. I need to lower the burden on the programmer.