```     inl Learning = Learning float32     inl player_pg {name actor learning_rate} cd =         inl action = Learning.RL.action {State Action final=Learning.RL.sampling_pg}
        inl input_size = Union.length_dense State
        inl num_actions = Union.length_one_hot Action

        inl net,_ =
            inl linear = Learning.Feedforward.linear
            Tuple.append (Tuple.wrap actor) (linear num_actions :: ())
            |> Learning.init cd input_size

        inl net =
            Union.mutable_function <| inl {state={state with net} input={input cd}} ->
                inl {action bck net} = action {net input} cd
                inl bck =
                    match state with
                    | {bck} -> List.cons (heap bck) bck
                    | _ -> List.singleton (heap bck)
                {state={net bck}; out=action}
            {state={net}; input={input=State; cd}}

        inl methods = {basic_methods with
            bet=inl s input -> s.data.net {input cd=s.data.cd}
            showdown=inl s v ->
                inl {state with net} = s.data.net.reset
                inl optimizer = Learning.Optimizer.sgd learning_rate
                Struct.iter (inl {optimize} -> optimize optimizer) net

                match state with
                | {bck} -> List.foldl (inl reward bck -> bck {reward} |> ignore; reward) (dyn (to float32 v)) bck |> ignore
                | _ -> ()
            game_over=inl s -> ()
            }

        Object
            .member_add methods
            .data_add {name; win=ref 0; net}
```

There is definitely a beauty to this now. Well, assuming that it works which is probably not the case since I haven't tested it yet. I am not sure I feel like it right now.

I just want to sit here and admire how the pg player looks for a bit now. In the last version I really had to think a lot on every turn whether to `indiv` and whether to `heap`. Coupled with never really being sure, it was like carrying a heavy anchor around. This code fragment here is a lot lighter in comparison despite them being close in the number of lines of code.

This code is just so crystal clear.

I am definitely done for the day. Hopefully I'll make some proper progress tomorrow. I was really eager yesterday when I went to bed early, but today turned out to be pretty standard after all.

This type sealing stuff is pretty difficult. It is one of the aspects of Spiral which I haven't quite yet mastered, but I will get it just like everything else.