"5:25pm.
```
The time it took to compile the Cuda kernels is: 00:00:01.8142592
Compiled the kernels into the following directory: C:\Users\Marko\Source\Repos\The Spiral Language\Temporary\bin\Release
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0

Unhandled Exception: System.Exception: impossible
```

```
                match state with
                | {bck} -> List.foldl (inl reward bck -> bck {reward} |> ignore; reward) (dyn (to float32 v)) bck |> ignore
                | _ -> failwith () "impossible"
```

I was suspecting this, but why?

5:25pm.

```
        inl {state=state' out} = match state() with () | _ as state -> f {state input}
        state := box ty state'
        out
```

Ah, now that I think about it, this is a bit weird. Shouldn't the state be boxed again?

...Ah, it is.

5:30pm. Why is this happening exactly?"