I got the new one to work.
The winrate is identical to the old one as it should be so I won't paste it here again.

```
    inl player_pg {name actor learning_rate} cd =
        inl action = Learning.RL.action {State Action final=Learning.RL.sampling_pg}

        inl input_size = Union.length_dense State
        inl num_actions = Union.length_one_hot Action

        inl net,_ =
            inl linear = Learning.Feedforward.linear
            Tuple.append (Tuple.wrap actor) (linear num_actions :: ())
            |> Learning.init cd input_size

        inl net =
            Union.mutable_function
                (inl {state={state with net} input={input cd}} ->
                    inl {action bck net} = action {net input} cd
                    inl bck =
                        match state with
                        | {bck=bck'} -> List.cons (heap bck) bck'
                        | _ -> List.singleton (heap bck)
                        |> dyn
                    {state={net bck}; out=action}
                    )
                {state={net}; input={input=State; cd}}

        inl methods = {basic_methods with
            bet=inl s input -> s.data.net {input cd=s.data.cd}
            showdown=inl s reward ->
                match s.data.net.reset with
                | {net bck} ->
                    inl reward = dyn (to float32 reward)
                    List.foldl (inl _ bck -> bck {reward}) () bck
                    inl optimizer = Learning.Optimizer.sgd learning_rate
                    Struct.iter (inl {optimize} -> optimize optimizer) net
                | _ -> ()
            game_over=inl s -> ()
            }

        Object
            .member_add methods
            .data_add {name; win=ref 0; net}
```

Here is the full working player once more for reference.

It is nice and clean. I am quite pleased by this.

Though it does not seem like much, it really just a short hop from this to an actor critic architecture. Now I have the function in this form I will have no trouble whatsoever adapting to it.

The first half of the function I should factor out, but it is not a big deal. I do not feel like doing it anymore today. Unless I am very unlucky, Without a doubt I will implement the Zap AC tomorrow.