Damn, I want to get this done. I really want to get this done. I went to bed earlier just so I can wake up and start programming earlier.
An AC player with Zap would be the first thing of value I will get in my 3.5 years of work.

Really, from the outside I must look like such a moron, to do all this work instead of taking the much easier route of using Python+PyTorch or Tensorflow.

It is going to start paying off here. There is no way Python can draw out the full power of ML in the way Spiral can. Maybe if the Julia guys take my views seriuosly that thing will become a competitive language, but that won't be anytime soon.

Damn, I am tired. I was tired for a long time.

All the goals and inspiration and the effort, just to make a little optimization. All of that to perfect the absolute basics. I've been a fool for way too long, this was always the right way from the start. It is impossible to fight while standing in a swamp.

I wish I could have done this work in the 00s instead of now, but in those days I was too busy trying to figure out omnipotence. Well, that has sure been figured out.

https://pathak22.github.io/large-scale-curiosity/

Well, let me read a little and then I will start. It is just another day. This won't go on forever. Eventually the project will be finished.

I have some new ideas for how to design a ML library. Just how network is now fused with state, I think I should also fuse the bck's with it. I am really gaining zero from having the layers like this. Rather than use those useless explicit lists, I should move more into the mutable_function.

This should be done now, not later when the changes will be hard to introduce.

I have a really good thing going with these mutable functions. I should follow up on this work.