```     inl run s input =         Struct.foldl_map (inl input {layer with apply} ->             inl input =                 inl input = {input}                 inl input = match layer with {weights} -> {input with weights} | _ -> input                 match layer with {state} -> {input with state} | _ -> input
            inl {x with out} = indiv join apply input s |> stack
            inl layer = match x with {bck} -> {layer with bck=heap bck} | _ -> layer
            inl layer = match x with {state} -> {layer with state=heap state} | _ -> layer
            layer, out
            ) input
```

Here is the new run function. It is much simplified. Let me also heap the state...Overwritten.

This is the direction I want to go in the library. Do everything with maps and folds - that is the wave of the future. Programming needs to be radically simplified to be tractable.

```
    inl init s =
        Struct.foldl_map (inl sublayer_size {x with init} ->
            inl {dsc size} = init sublayer_size
            inl weights = initialize s dsc |> heap
            inl optimize =
                match x with
                | {optimize} optimizer -> optimize (optimizer s)
                | _ optimizer -> Struct.iter (optimizer s) weights
            {x without init with optimize weights}, size
            )
```

Let me also change this function. Given that I am using PRONG and will be using Zap and who knows what else in the future, it does not make sense to keep passing the damn SGD. I'll just do more in the layers.